+++ src/app/studyDetail/studyContent/modal/appliedModal/AppliedModalContainer.tsx
