+++ src/app/studyDetail/studyContent/modal/appliedModal/AppliedModalPresenter.tsx
