+++ src/app/studyDetail/studyContent/modal/appliedModal/index.scss
