+++ src/app/studyDetail/studyContent/modal/applyCancelModal/ApplyCancelModalContainer.tsx
