+++ src/app/studyDetail/studyContent/modal/applyCancelModal/ApplyCancelModalPresenter.tsx
