+++ src/app/studyDetail/studyContent/modal/applyModal/ApplyModalContainer.tsx
