+++ src/app/studyDetail/studyContent/modal/applyModal/ApplyModalPresenter.tsx
