+++ src/app/studyDetail/studyContent/modal/applyModal/index.scss
