+++ src/app/studyDetail/studyContent/modal/approveCancelModal/ApproveCancelModalPresenter.tsx
